+++ app/src/components/Header.vue
@@ -15,27 +15,30 @@ const emit = defineEmits(["toggle-menu"]);
 
 <template>
   <div>
-    <Search id="search-box" />
-
     <div id="header-bar" class="border-b border-gray-400 bg-gray-300 py-2">
       <ul
         class="flex flex-row justify-between font-medium text-black sm:text-lg md:text-xl"
       >
-        <div class="ml-6 flex flex-row space-x-2">
+        <div
+          class="ml-6 flex flex-row content-center space-x-2 sm:text-sm lg:text-base"
+        >
           <div
-            class="mt-[0.125rem] hover:cursor-pointer hover:bg-gray-400"
+            class="m-auto h-fit hover:cursor-pointer hover:bg-gray-400"
             @click="emit('toggle-menu')"
           >
             <hamburger-menu />
           </div>
-          <li id="header-title">
+          <li id="header-title" class="m-auto h-fit">
             {{ title }}
             <template v-if="version != ''">({{ version }})</template>
           </li>
         </div>
+
+        <Search />
+
         <div
           id="header-links"
-          class="mr-6 flex flex-row space-x-6 text-base text-gray-600"
+          class="mr-6 flex flex-row space-x-6 text-gray-600 sm:text-sm lg:text-base"
         >
           <a
             v-for="link in links"